lunatik: the percpu object gets files of its own
lunatik_core.c held the runtime and, since the percpu object, the class
that holds one runtime per CPU id too, and lunatik.h its type, dispatch
pin and accessors among the runtime's. lunatik_percpu.c takes the
class, its constructor and its teardown; lunatik_percpu.h takes the
declarations, included at the end of lunatik.h as lunatik_val.h is,
since lunatik_run is a macro and resolves lunatik_pin where it is used.
The core lends the pieces both constructors share, lunatik_newruntime
and the context option, through lunatik.h. Nothing changes behaviour.
