+++ Kbuild
@@ -33,7 +33,7 @@ lunatik-objs += lua/lapi.o lua/lcode.o lua/lctype.o lua/ldebug.o lua/ldo.o \
 	lua/lcorolib.o lua/ldblib.o lua/lstrlib.o \
 	lua/ltablib.o lua/lutf8lib.o lua/lmathlib.o lua/liolib.o lua/linit.o \
 	lua/loadlib.o $(KLIBC_USR)/klibc/arch/$(KLIBC_ARCH)/setjmp.o \
-	lunatik_aux.o lunatik_obj.o lunatik_val.o lunatik_core.o
+	lunatik_aux.o lunatik_obj.o lunatik_val.o lunatik_core.o lunatik_percpu.o
 
 ifeq ($(CONFIG_64BIT),)
 lunatik-objs += $(KLIBC_LIBGCC)/__udivmoddi4.o	\
