+++ lunatik_core.c
@@ -163,8 +163,6 @@ static int lunatik_cpu(lua_State *L)
 	return 1;
 }
 
-static int lunatik_percpu(lua_State *L);
-
 static const luaL_Reg lunatik_lib[] = {
 	{"runtime", lunatik_lruntime},
 	{"percpu", lunatik_percpu},
@@ -237,7 +235,7 @@ static int lunatik_runscript(lua_State *L)
 	return 1; /* callback */
 }
 
-static int lunatik_newruntime(lunatik_object_t **pruntime, lua_State *Lfrom, const char *script, lunatik_opt_t opt, int cpu)
+int lunatik_newruntime(lunatik_object_t **pruntime, lua_State *Lfrom, const char *script, lunatik_opt_t opt, int cpu)
 {
 	lunatik_object_t *runtime;
 	lua_State *L;
@@ -288,14 +286,6 @@ int lunatik_runtime(lunatik_object_t **pruntime, const char *script, lunatik_opt
 }
 EXPORT_SYMBOL(lunatik_runtime);
 
-static inline lunatik_opt_t lunatik_checkcontext(lua_State *L, int ix)
-{
-	static const char *const contexts[] = {"process", "softirq", "hardirq", NULL};
-	static const lunatik_opt_t opts[] = {LUNATIK_OPT_NONE, LUNATIK_OPT_SOFTIRQ, LUNATIK_OPT_HARDIRQ};
-
-	return opts[luaL_checkoption(L, ix, "process", contexts)];
-}
-
 /***
 * Creates a new Lunatik runtime executing the given script.
 * @function runtime
@@ -318,99 +308,6 @@ static int lunatik_lruntime(lua_State *L)
 	if (lunatik_newruntime(pruntime, L, script, opt, LUNATIK_CPU_NONE) != 0)
 		lua_error(L);
 	lunatik_setclass(L, &lunatik_class, true);
-	return 1;
-}
-
-/***
-* Set of runtimes, one per CPU id, running the same script. A callback dispatched
-* through it reaches the instance of the CPU it fired on.
-* @type percpu
-*/
-static void lunatik_releasepercpu(void *private)
-{
-	lunatik_object_t * __percpu *runtimes = lunatik_percpuruntimes(private);
-	int cpu;
-
-	for_each_possible_cpu(cpu) {
-		lunatik_object_t *runtime = *per_cpu_ptr(runtimes, cpu);
-		if (runtime != NULL) /* last reference: a stop would lock, and this can run in softirq */
-			lunatik_putobject(runtime);
-	}
-	free_percpu(runtimes);
-}
-
-static const lunatik_class_t lunatik_percpu_class;
-
-static inline lunatik_object_t * __percpu *lunatik_checkruntimes(lua_State *L, int ix)
-{
-	lunatik_object_t *object = lunatik_checkobject(L, ix);
-	lunatik_argcheckclass(L, ix, object, &lunatik_percpu_class);
-	return lunatik_percpuruntimes(object->private);
-}
-
-/***
-* Closes every instance, releasing their Lua states.
-* @function stop
-*/
-static int lunatik_stoppercpu(lua_State *L)
-{
-	lunatik_object_t * __percpu *runtimes = lunatik_checkruntimes(L, 1);
-	int cpu;
-
-	for_each_possible_cpu(cpu) {
-		lunatik_object_t *runtime = *per_cpu_ptr(runtimes, cpu);
-		if (runtime != NULL)
-			lunatik_closeprivate(runtime);
-	}
-	return 0;
-}
-
-static const luaL_Reg lunatik_percpu_mt[] = {
-	{"__gc", lunatik_deleteobject},
-	{"__close", lunatik_stoppercpu},
-	{"stop", lunatik_stoppercpu},
-	{NULL, NULL}
-};
-
-static const lunatik_class_t lunatik_percpu_class = {
-	.name = "percpu",
-	.methods = lunatik_percpu_mt,
-	.release = lunatik_releasepercpu,
-	.opener = luaopen_lunatik,
-	.opt = LUNATIK_OPT_PERCPU | LUNATIK_OPT_EXTERNAL,
-};
-
-/***
-* Creates one runtime per CPU id, each loading the given script in the calling context.
-* The instances are dispatched by CPU: see `lunatik.cpu` and `runner.run`.
-* @function percpu
-* @tparam string script script name (e.g., `"mymod"` loads `/lib/modules/lua/mymod.lua`)
-* @tparam[opt="process"] string context execution context, as in `lunatik.runtime`
-* @treturn percpu
-* @raise if allocation fails or the script errors on load, after releasing the instances
-*   already created
-* @within lunatik
-*/
-static int lunatik_percpu(lua_State *L)
-{
-	const char *script = luaL_checkstring(L, 1);
-	lunatik_opt_t opt = lunatik_checkcontext(L, 2);
-	int cpu;
-
-	lunatik_object_t *object = lunatik_newobject(L, &lunatik_percpu_class, 0, opt);
-	lunatik_object_t * __percpu *runtimes = alloc_percpu(lunatik_object_t *);
-
-	if (runtimes == NULL)
-		lunatik_enomem(L);
-	object->private = runtimes;
-
-	for_each_possible_cpu(cpu) {
-		if (lunatik_newruntime(per_cpu_ptr(runtimes, cpu), L, script, opt, cpu) != 0) {
-			object->private = NULL;
-			lunatik_releasepercpu(runtimes); /* release the instances now, not on collection */
-			lua_error(L);
-		}
-	}
 	return 1;
 }
